Susan Polgar versus Karpov Drawn
This is a game that got drawn, but also containts a lot pieces that are the same that can end of on the same square. Like Rcc8 has to be written because another R can move to c8 that is on another file. This is also the first game with the year 2000 and time control (all spaces filled out at the top). This goes all the way down the page.